requirements.txt
altgraph==0.17
cffi==1.14.0
fbs==0.8.4
future==0.18.2
numpy==1.18.1
opencv-contrib-python==4.2.0.32
pefile==2019.4.18
Pillow==7.0.0xia
pycparser==2.19
PyInstaller==3.6
PyQt5==5.9.2
pywin32-ctypes==0.2.0
QtPy==1.9.0
sip==4.19.8
